+++ nuxeo/client.py
@@ -40,7 +40,7 @@ from .constants import (
     TIMEOUT_READ,
 )
 from .exceptions import BadQuery, Forbidden, HTTPError, Unauthorized
-from .utils import json_helper
+from .utils import get_response_content, json_helper
 
 try:
     from typing import TYPE_CHECKING
@@ -259,11 +259,16 @@ class NuxeoClient(object):
         # to set `default` to `None`.
         default = kwargs.pop("default", object)
 
+        _kwargs = {k: v for k, v in kwargs.items() if k != "params"}
         logger.debug(
-            ("Calling {!r} with headers={!r}, params={!r} and cookies={!r}").format(
+            (
+                "Calling {} {!r} with headers={!r}, params={!r}, kwargs={!r} and cookies={!r}"
+            ).format(
+                method,
                 url,
                 headers,
                 kwargs.get("params", data if not raw else {}),
+                _kwargs,
                 self._session.cookies,
             )
         )
@@ -273,13 +278,12 @@ class NuxeoClient(object):
             resp = self._session.request(
                 method, url, headers=headers, auth=self.auth, data=data, **kwargs
             )
+            self._log_response(resp)
             resp.raise_for_status()
         except Exception as exc:
             if default is object:
                 raise self._handle_error(exc)
             resp = default
-        else:
-            self._log_response(resp)
         finally:
             # Explicitly break a reference cycle
             exc = None
@@ -359,7 +363,13 @@ class NuxeoClient(object):
         if force or self._server_info is None:
             response = self.request("GET", "json/cmis", default={})
             if isinstance(response, requests.Response):
-                info = response.json()["default"]
+                try:
+                    info = response.json()["default"]
+                except Exception:
+                    logger.warning(
+                        "Invalid response data when called server_info()", exc_info=True
+                    )
+                    info = None
             else:
                 info = response
             self._server_info = info
@@ -412,6 +422,9 @@ class NuxeoClient(object):
         content_size = int(headers.get("content-length", 0))
         chunked = headers.get("transfer-encoding", "") == "chunked"
 
+        if response.status_code and response.status_code >= 400:
+            # This is a request ending on an error
+            content = get_response_content(response, limit_size)
         if response.url.endswith("site/automation"):
             # This endpoint returns too many information and pollute logs.
             # Besides contents of this call are stored into the .operations attr.
@@ -431,22 +444,7 @@ class NuxeoClient(object):
             content = "<binary data ({:,} bytes)>".format(content_size)
         elif chunked or content_size > 0:
             # At this point, we should only have text data not bigger than *limit_size*.
-            # Do not use response.text as it will load the chardet module and its
-            # heavy encoding detection mecanism. The server will only return UTF-8.
-            # See https://stackoverflow.com/a/24656254/1117028 and NXPY-100.
-            try:
-                content = response.content.decode("utf-8", errors="replace")
-                content_size = len(content)
-                if content_size > limit_size:
-                    content = content[:limit_size]
-                    content = "{} [...] ({:,} bytes skipped)".format(
-                        content, content_size - limit_size
-                    )
-            except (MemoryError, OverflowError):
-                # OverflowError: join() result is too long (in requests.models.content)
-                # Still check for memory errors, this should never happen; or else,
-                # it should be painless.
-                content = "<no enough memory ({:,} bytes)>".format(content_size)
+            content = get_response_content(response, limit_size)
         else:
             # response.content is empty when *void_op* is True,
             # meaning we do not want to get back what we sent
@@ -454,8 +452,8 @@ class NuxeoClient(object):
             content = "<no content>"
 
         logger.debug(
-            "Response from {!r}: {!r} with headers {!r} and cookies {!r}".format(
-                response.url, content, headers, response.cookies
+            "Response from {!r} [{}]: {!r} with headers {!r} and cookies {!r}".format(
+                response.url, response.status_code, content, headers, response.cookies
             )
         )
 
@@ -510,8 +508,8 @@ class Nuxeo(object):
 
     def __repr__(self):
         # type: () -> Text
-        fmt = "{name}<client={cls.client!r}>"
-        return fmt.format(name=type(self).__name__, cls=self)
+        fmt = "{name}<version={ver!r}, client={cls.client!r}>"
+        return fmt.format(name=type(self).__name__, cls=self, ver=__version__)
 
     def __str__(self):
         # type: () -> Text